// angular-client/akira-ng/src/app/model/response/open-library-book-response.model.ts
export interface OpenLibraryBookResponse {
  cover_i?: number;
  has_fulltext?: boolean;
  edition_count?: number;
  title: string;
  author_name?: string[];
  first_publish_year?: number;
  key: string;
  ia?: string[];
  author_key?: string[];
  public_scan_b?: boolean;
}

export interface PagedOpenLibraryResponse {
  start: number;
  num_found: number;
  docs: OpenLibraryBookResponse[];
}